Moved and renamed the haplotype scorer code.
